notebook/requirements/test.txt
-r common.txt
pytest
flake8
isort==4.3.21
pytest-django
mixer
pytest-cov
pytest-pythonpath